lib: fwts_acpi: rsdp requires a 64 bit xsdt

+++ src/lib/include/fwts_acpi.h
@@ -122,7 +122,7 @@ typedef struct {
 	uint8_t		revision;
 	uint32_t	rsdt_address;
 	uint32_t	length;
-	uint32_t	xsdt_address;
+	uint64_t	xsdt_address;
 	uint8_t		extended_checksum;
 	uint8_t		reserved[3];
 } __attribute__ ((packed)) fwts_acpi_table_rsdp;
